Init templates and separate the first template: FieldTemplate
