+++ packages/react-jsonschema-form-bootstrap/src/components/templates/index.js
@@ -1,0 +1,3 @@
+import FieldTemplate from './FieldTemplate';
+
+export default { FieldTemplate };
