+++ packages/react-jsonschema-form-bootstrap/src/index.js
@@ -1,3 +1,4 @@
+import templates from './components/templates';
 import widgets from './components/widgets';
 
-export default { widgets };
+export default { templates, widgets };
