+++ packages/react-jsonschema-form/src/components/Form.js
@@ -6,9 +6,16 @@ import baseFields from './fields';
 
 class Form extends Component {
   render() {
-    const { theme, fields = {}, widgets = {}, ...props } = this.props;
+    const {
+      theme,
+      fields = {},
+      templates = {},
+      widgets = {},
+      ...props
+    } = this.props;
     const components = {
       fields: { ...baseFields, ...fields },
+      templates: { ...theme.templates, ...templates },
       widgets: { ...theme.widgets, ...widgets }
     };
 
@@ -19,6 +26,9 @@ class Form extends Component {
 if (process.env.NODE_ENV !== 'production') {
   Form.propTypes = {
     theme: PropTypes.shape({
+      templates: PropTypes.objectOf(
+        PropTypes.oneOfType([PropTypes.func, PropTypes.object])
+      ).isRequired,
       widgets: PropTypes.objectOf(
         PropTypes.oneOfType([PropTypes.func, PropTypes.object])
       ).isRequired
